requirements.txt.txt
alembic           1.14.0
bcrypt            4.2.0
blinker           1.9.0
click             8.1.7
colorama          0.4.6
dnspython         2.7.0
email_validator   2.2.0
Flask             3.0.3
Flask-Bcrypt      1.0.1
Flask-Login       0.6.3
Flask-Mail        0.10.0
Flask-Migrate     4.0.7
Flask-SQLAlchemy  3.1.1
Flask-WTF         1.2.2
idna              3.10
itsdangerous      2.2.0
Jinja2            3.1.4
Mako              1.3.6
MarkupSafe        3.0.2
pip               24.2
pydeps            2.0.1
python-dotenv     1.0.1
SQLAlchemy        2.0.36
stdlib-list       0.11.0
typing_extensions 4.12.2
Werkzeug          3.1.3
WTForms 